Simple implementation of the secpolicy functions
This fixes a security hole that allows unprivileged users to perform
actions that upstream illumos usually reserves for root except when a
specific privilege has been granted to a non-root user.

To tighten up security on OS X, we now enforce root privileges for most
read-write operations, while still allowing unprivileged users read-only
access. Even with the tightened security, this is still a marked
improvement in convenience compared to the current ZFS on Linux approach
of requiring root access for all ZFS interactions, including those that
are read only. Our more permissive approach may in actual practice be
safer on balance since it encourages users only to resort to sudo when
absolutely necessary, as opposed to requiring them to use it in all
cases, in which case the user will never be advised that any particular
action requires root access to proceed since the user will always be
operating with root access.

Without secpolicy but with privilege escalation, we've had the similar
problem of "too much" convenience, which not only permitted unprivileged
users to perform relatively benign read-write operations such as
creating a new dataset, but also permitted unprivileged users to perform
destructive read-write operations such as destroying existing datasets.
Although this allowed for maximum convenience, it is not secure,
especially on a multiuser system.

The original intent of allowing kernel-side privilege escalation for
unprivileged users when using ZFS was to resolve a bug that caused a
hang when zpool scrub was invoked by a user without root privileges.
While this escalation does work around that specific problem, it opens
up a security hole that this commit patches. This commit implicitly
prevents the problem of a hang upon scrub invocation by an unprivileged
user by explicitly denying that command before it gets to the point of
hanging, as is demonstrated in the following output:

bash-4.3$ zpool scrub internal
cannot scrub internal: permission denied
bash-4.3$ sudo zpool scrub internal
Password:
bash-4.3$ zpool status
  pool: internal
 state: ONLINE
  scan: scrub in progress since Wed Mar 18 00:57:39 2015
    924M scanned out of 26.2G at 185M/s, 0h2m to go
    0 repaired, 3.44% done
config:

	NAME                                 STATE     READ WRITE CKSUM
	internal                             ONLINE       0     0     0
	  APPLE_SSD_SM512E-S0VMNYAC538462:4  ONLINE       0     0     0

errors: No known data errors