Rework debug cmn_err() API
Previously messages to cmn_err would either go nowhere, or be sent
to the serial console. (Not frequently used on OSX).

Now we print them to the system.log with the priority, this enables
zfs_recover to be set, and we will still receive message when we do
not call panic().