smd: spl side of dynamic arc_c_min adjustment
When vmem_update notices that memory is very low, it will
cause arc_shrink() (in zfs) to reduce arc_c_min;
conversely, when there is ample memory, it will cause an
increase in arc_c_min up to the originally saved value.

The originally saved value is set in arc_init(); it can be
updated by sysctl.