faster large block allocation
Getting large blocks from XNU faster means that we don't need as large
an initial allocation in vmem_init().    That large allocation is often
very sparse anyway, since almost all allocations are <= 1MiB unless there
are active datasets with recordsize >= 2MiB.

(zvols with volblocksize > 128k are presently forbidden).

recordsize > 1 MiB is terrible for random access anyway, and even the
much smaller initial allocation provides ample space for large records.

Additionally, when we are streaming, the lifetime of the large blocks in
ARC that are not re-zfetched or redirtied will be small, so if we cannot
satisfy the 2MiB or larger allocation right away, a half-second wait for
free space to appear before asking xnu for two (or one, in tight memory)
blocks is reasonable.

(We ask for two so we can put another onto the freelist in anticipation
of zfetch or the like needing an allocation of the next size.   Soon
enough a vmem_walk of spl_root_arena or a free of the initial allocation
will return it to xnu if it's not needed.)

If we cannot get two * size because of memory pressure, then we try
size, and if we cannot get that then we fail if that is what the
allocator wants, or we loop around waiting until pressure drops enough
that we can ask xnu to allocate one or two times size, or until a free
block of the correct size appears.