don't double small allocs & minor printf change
Somewhat academic as the previous commit practically never winds up
going to xnu for an allocation

(after 15 602 spl_root_arena.parent_alloc, there have been only
41 spl_vmem_large_alloc, all of which were satisfied from the initial
large allocation; spl_root_arena_parent.parent_xalloc (the fall-through)
is 1718, spl_root_arena_parent.populate_wait (the for loop) is 48 525,
and the spl_root_arena_parent.fail (second pass through wait loop) and
spl_root_arena_parent.mem_import (successful call to the XNU allocator)
are both zero -- IOW so far there would have been no small-block WOAH anyway.)

but if a small alloc happens to be unfulfilled at the end of the for
loop, it would be good to know about it in the log.

Additionally, if a small alloc goes to vmem_add_or_return_...() then a
second small span should not be added.   (It's good to add a second
large span if possible; if a sub 1 MiB alloc is needed again in short
order, it should go back to waiting in hopes of a free elsewhere or of a
population by the refill thread.)