rework initial large reserve space
The initial large reserve space is awkward to inspect and
is under-used.

This fixes the first half of that and also provides use
by arbitrarily-small blocks when the xnu is scanning for
pages.

spl_root_arena needs a bit of memory to start with, we'll take 32 MiB from
the reserve for its vmem_create.