heap_parent is redundant, also update big comment
The vmem_metadata is allocated sufficiently early that parenting it to
the reserve arena saves some early activity (and very likely several
calls to the XNU allocator) compared to running it through
spl_root_arena.  It is not very busy; its children will ask it for about
6000 allocations, and frees will be about 17 through the whole system
uptime, and the amount of allocations it makes to its parent is less
than two hundred over the whole system uptime.  It's also not very
large, and should be less than 32 MiB, which is a small fraction of the
reserve arena.