create timed/conditional allocator functions
new spl_root_allocator()

update kstats, attempt recover from free_arena

rip out spl_refill_thread_request mechanism
- --
Don't do a recursive vmem_alloc

This is probably bad :

spl_root_allocator(...) -> vmem_allocate(spl_root, ...)

even though we can probably fudge it because the first call has its vmp
as spl_root_allocator_parent, and at least superficially it appears to work.

The problem is that while we go from the freelist to the non-freelist
with the vmem_allocate, it is not clear that the second allocation will
not cause problems when vmem_xfree() is triggered on it.

It's easier just to avoid recursing, as we can just "fail" and return(NULL)
which ends up here :
...
} else if (vmem_canalloc(vmp, aneeded)) {
/*
 * Our import failed, but another thread
 * added sufficient free memory to the arena
 * to satisfy our request.  Go back and
 * grab it.
 */
   ASSERT(resv >= VMEM_SEGS_PER_MIDDLE_ALLOC);
   goto do_alloc;
 }
- --
avoid totally filling reserve

Symptom:

vmem_xalloc: vmem waiting for 131072 sized alloc for spl_large_reserve_arena

This is a waiting vmem_metadata_arena; we probably don't want that, even
if it's rare and vmem_metadata_arena children can cope with this.
- --
xnu_import_arena for use by spl_root_allocator.
Avoid having a non-NULL ffunc for xnu_import_arena,
so we have to vacuum occasionally.

force allocate if we've been looping for a long time