shorten wait times to improve throughput
cv_timedwait_hires() -> msleep() takes nanoseconds.

the "res" argument is not useful, and 0 is acceptable.

shorter wait times mean less latency when spl_root_arena is low on free
segments, and since throughput is latency-limited, big writes and long
timeouts can be annoyingly slow on fast media, so long as there is
allocatable space in any of the three arenas when spl_root_allocator()
is called.

this is a direct trade-off with faster acquisition of memory, which can
push the system into poor behaviour (userland compressing, maybe
pageouts, and possibly glitches).

the cv_timedwait_hires calls often proceed to timeout
(but this contrasts with the cv_wait() in vmem_xalloc())

motivation: previous e-powering the fill thread means the
cv_timedwait_hires() calls can be made frequently when arc is big, which
on a spinny-disk pool with writers and a streaming reader can lead to
read stalls, which is no good.  ideally typical-case wait times through
spl_root_allocator() will be about the same as the worst-case *random* I/O
latency of the underlying storage, so the waits on the large reserve
arena and for freeing memory in the arenas might need further shrinking
in the face of very fast storage.

the worst-case wait can be arbitrarily long (again, it is cv_wait() in
illumos); the only problem is that after about seconds (!) zfs gets
stuck mainly in spa mutexes.    spl_root_allocator() will never allow
that to happen, as it will always resort to a direct allocation request
to xnu (which will always grant it, or panic the system) within a couple
of seconds.