monitoring scripts that flash away in big iterm windows
these are the most interesting and relevant kstats, just
for interest and to revision-control them here.

splwatch+splstat : watch the most important stats,
covering the amount of activity done with xnu's allocator
and freer, and the paths taken to get to each.

ziowatch : watch the size and activity of the two zio
arenas bs xnu activity and calculated spl_free.

splwatch2: watch the two source arenas, both the static
large block obtained in vmem_init() and the dynamic import
and release arena which handles; it also watches the
relevant bits of spl_root_arena (alloc/free = number of times a
child arena asked for or returned memory; these will mostly be qcaches
asking for 256k, or bigger allocs; parent_alloc = number
of times a "source" has deposited memory into
spl_root_arena (this can be the refill thread or recovery
of memory from the free_arena); parent_free = number of
deposits of memory into the free_arena (these will be
span-sized, typically 1MiB or larger).