deal with sequential stutter in extreme corner
We can be in a condition where spl_free is large enough to keep
arc_no_grow FALSE for various reasons, partly because in low memory
the XNU allocator is slow compared to the memory-scavenging that
XNU does (scan-based compression or eviction, for example) and
partly because we are likely to keep having to go to
spa_root_allocator() -> timed_xnu_alloc_root_xnu(), which can
also be slow.   This is especially the case where recordsizes
are large compared to spl_minalloc (e.g. >= 1 MiB) and compared
to zfetch_array_rd_sz (also >= 1 MiB).

Although the immediate solution is to manually send pressure:

    	sysctl -w kstat.spl.misc.spl_misc.spl_spl_free_fast_pressure=1
	sysctl -w kstat.spl.misc.spl_misc.spl_spl_free_manual_pressure=$a

where $a is not too far from 1 (it's in units of megabytes), and
the problem can be completely avoided by setting an arc_c_max that is
low compared kstat.spl.misc.spl_misc.root_refill_high_percent (e.g., if
that's 60 (default), then set arc_c_max to 55), and is unlikely to hit
when the streaming reader is doing small reads compared to
zfetch_array_rd_sz, or at all on datasets with recordsizes near 128k.

This patch signals pressure to arc when spa_root_allocator has been at
least a few milliseconds and its first attempt to ask XNU for a chunk of
memory was not made because there weren't enough clearly free pages to
satisfy the request.