take into account free memory we have from xnu
If we have "lots" of memory -- namely if we can allocate a 64 MiB
segment in the reserve arena, or have 128 MiB free in the reserve arena,
or have 256 MiB free in the root arena, then we don't want to react
strongly to transient system pressure.  On a busy system, this leads to
an ARC that is rarely far from arc_c_min, plus a wildly underused
reserve arena, plus probably a wildly underused root arena (the fill
thread may still be topping up the root arena, which ARC does not then
consume because of the pressure signals sent by spl_free_thread()).

The pressure signals are fast and strong enough that in well under a
second GiB can be freed back to the system in the presence of real
pressure, like when one fires up a browser that opens many windows and
tabs as it starts.