Remove mountroot hijack, handle NULL getrootdir
With 10.12, we no longer need to hijack mountroot to enable ZFS Booting,
this means we will only support ZFS Booting from 10.12 (as we are allowed
to handle mountroot by Apple). ZFS Booting is a new feature, so we are
not removing an existing feature.

ZVOL Booting does not use mountroot hijack.

getdrootdir() which calls vfs_rootvnode() will panic if called before
root is mounted, so we need to wrap this call with a delay switch.
It is enabled from VFS_ROOT, when we are booting / using mountroot.
Or it is enabled in zfs_osx init, if it detects we are not booting.

This could do with a bit more cleanup.